+++ Poly/src/Poly/Rendering/SceneRenderer.cpp
@@ -1,16 +1,9 @@
 #include "SceneRenderer.h"
 
 #include "Poly/Model/Mesh.h"
-#include "Poly/Model/Material.h"
-#include "Poly/Core/RenderAPI.h"
 #include "Platform/API/Buffer.h"
-#include "Platform/API/Texture.h"
-#include "Platform/API/Sampler.h"
-#include "Platform/API/DescriptorSet.h"
 #include "Poly/Rendering/RenderScene.h"
 #include "Poly/Rendering/RenderGraph/RenderContext.h"
-#include "Poly/Rendering/RenderGraph/RenderGraphProgram.h"
-#include "Poly/Resources/ResourceManager.h"
 
 namespace Poly
 {
@@ -24,7 +17,7 @@ namespace Poly
 		CommandBuffer* commandBuffer = context.GetCommandBuffer();
 
 		const SceneBatch* pSceneBatch = context.GetSceneBatch();
-		uint32 instanceCount = pSceneBatch->Matrices.size();
+		uint32 instanceCount = static_cast<uint32>(pSceneBatch->Matrices.size());
 
 		Ref<Mesh> pMesh = pSceneBatch->MeshInstance.pMesh;
 		const Buffer* pIndexBuffer = pMesh->GetIndexBuffer();
